+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGES:
 * Added a new script iam/show-password-policy.py. ([@TGWolf][])
 * Added a new section for ssm scripts. ([@TGWolf][])
 * Added a new script ssm/list-ssm-parameters.py. ([@TGWolf][])
+* Added a new script ssm/get-ssm-parameter.py. ([@TGWolf][])
 
 ## 1.0.6 (November 22, 2018)